Refactor: 화살표 아이콘 공용 에셋으로 통일 (#154)
