+++ src/features/main/ui/GenreSection.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import styled from "styled-components";
 import { useNavigate } from "react-router-dom";
 import ContentCard from "../../../shared/components/ContentCard";
-import RightArrow from "../../../shared/assets/icons/main/rightArrow.svg?react";
+import RightArrow from "../../../shared/assets/icons/common/arrow/rightArrow.svg?react";
 import Emotional from "../../../shared/assets/icons/common/genreIcon/emotional.svg?react";
 import Horror from "../../../shared/assets/icons/common/genreIcon/horror.svg?react";
 import Detective from "../../../shared/assets/icons/common/genreIcon/detective.svg?react";
@@ -187,15 +187,15 @@ const LocationSearchButton = styled.p`
 
 const StyledRightArrow = styled(RightArrow)`
   margin-left: 0.75rem;
-  width: 0.5rem;
+  width: 0.9375rem;
+  height: 0.9375rem;
+  fill: var(--RIU_Monochrome-500, #515467);
 
   @media (max-width: 1024px) {
     margin-left: 0.46875rem;
-    width: 0.9375rem;
   }
   @media (max-width: 768px) {
     margin-left: 0.125rem;
-    width: 0.9375rem;
   }
 `;
