+++ src/features/main/ui/LevelSection.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import styled from "styled-components";
 import { useNavigate } from "react-router-dom";
 import ContentCard from "../../../shared/components/ContentCard";
-import RightArrow from "../../../shared/assets/icons/main/rightArrow.svg?react";
+import RightArrow from "../../../shared/assets/icons/common/arrow/rightArrow.svg?react";
 import Level1Icon from "../../../shared/assets/icons/common/levelIcon/level1.svg?react";
 import Level2Icon from "../../../shared/assets/icons/common/levelIcon/level2.svg?react";
 import Level3Icon from "../../../shared/assets/icons/common/levelIcon/level3.svg?react";
@@ -179,15 +179,15 @@ const LocationSearchButton = styled.p`
 
 const StyledRightArrow = styled(RightArrow)`
   margin-left: 0.75rem;
-  width: 0.5rem;
+  width: 0.9375rem;
+  height: 0.9375rem;
+  fill: var(--RIU_Monochrome-500, #515467);
 
   @media (max-width: 1024px) {
     margin-left: 0.46875rem;
-    width: 0.9375rem;
   }
   @media (max-width: 768px) {
     margin-left: 0.125rem;
-    width: 0.9375rem;
   }
 `;
