+++ src/features/themeDetail/ui/ThemeOverviewCard.jsx
@@ -4,9 +4,9 @@ import { useNavigate } from "react-router-dom";
 import ThumbnailImg from '../../../shared/assets/images/common/thumbnailImg.png';
 import InfoIcon from '../../../shared/assets/icons/themeDetail/infoIcon.svg?react';
 import ShareIcon from '../../../shared/assets/icons/themeDetail/shareIcon.svg?react';
-import HeartIcon from '../../../shared/assets/icons/common/heart_default.svg?react';
-import HeartIcon2 from '../../../shared/assets/icons/common/heart_hover.svg?react';
-import HeartIcon3 from '../../../shared/assets/icons/common/heart_active.svg?react';
+import HeartIcon from '../../../shared/assets/icons/common/heart/heart_default.svg?react';
+import HeartIcon2 from '../../../shared/assets/icons/common/heart/heart_hover.svg?react';
+import HeartIcon3 from '../../../shared/assets/icons/common/heart/heart_active.svg?react';
 import ReservationIcon from '../../../shared/assets/icons/themeDetail/reservationIcon.svg?react';
 import ScheduleIcon from '../../../shared/assets/icons/themeDetail/scheduleIcon.svg?react';
 import PropTypes from 'prop-types';
