+++ src/features/themeDetail/ui/ThemeReviewSection.jsx
@@ -2,8 +2,8 @@ import styled from "styled-components";
 import { useState, useEffect } from "react";
 import RatingStar from "../../../shared/assets/icons/themeDetail/ratingStar.svg?react";
 import ThemeReview from "./ThemeReview";
-import RightArrowIcon from "../../../shared/assets/icons/survey/rightArrowIcon.svg?react";
-import LeftArrowIcon from "../../../shared/assets/icons/survey/leftArrowIcon.svg?react";
+import RightArrowIcon from "../../../shared/assets/icons/common/arrow/rightArrow.svg?react";
+import LeftArrowIcon from "../../../shared/assets/icons/common/arrow/leftArrow.svg?react";
 import NoDataIcon from "../../../shared/assets/images/common/noData/noDataImageSmall.png";
 import PencilIcon from "../../../shared/assets/icons/themeDetail/pencilIcon.svg?react";
 import { getThemeReviewsListAPI } from "../api/themeDetailAPI";
