+++ src/pages/myPage/ReservationsPage.jsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import { useNavigate } from "react-router-dom";
-import LeftArrow from "../../shared/assets/icons/myPage/leftArrow.svg?react";
+import LeftArrow from "../../shared/assets/icons/common/arrow/leftArrow.svg?react";
 import useDevice from "../../shared/hooks/useDevice";
 import CalendarSection from "../../features/mypage/ui/reservations/CalendarSection";
 import ScheduleSection from "../../features/mypage/ui/reservations/ScheduleSection";
@@ -76,6 +76,10 @@ const BackButtonWrapper = styled.div`
 const StyledLeftArrow = styled(LeftArrow)`
   width: 1.25rem;
   height: 1.25rem;
+
+  path {
+    fill: var(--RIU_Monochrome-200, #717486);
+  }
 `;
 
 const BackButtonText = styled.div`
