+++ src/shared/components/ContentCard.jsx
@@ -5,9 +5,9 @@ import PropTypes from 'prop-types';
 import ThumbnailImg from '../assets/images/common/thumbnailImg.png';
 import AwardsIcon from '../assets/icons/common/awards.svg?react';
 import CautionIcon from '../assets/icons/common/cautionIcon.svg?react';
-import HeartIcon from '../assets/icons/common/heart_default.svg?react';
-import HeartIcon2 from '../assets/icons/common/heart_hover.svg?react';
-import HeartIcon3 from '../assets/icons/common/heart_active.svg?react';
+import HeartIcon from '../assets/icons/common/heart/heart_default.svg?react';
+import HeartIcon2 from '../assets/icons/common/heart/heart_hover.svg?react';
+import HeartIcon3 from '../assets/icons/common/heart/heart_active.svg?react';
 import { formatNumberWithCommas } from '../utils/formatUtils';
 import { levelTextConversion, genreListConversion, satisfactionConversion } from '../utils/dataUtils';
 import useDevice from '../hooks/useDevice';
