+++ src/shared/components/filter/RegionFilter.jsx
@@ -8,7 +8,7 @@ import {
   DropDownIcon,
   MenuWrapper } from "../../styles/FilterStyles.js";
 import Location from '../../assets/icons/genre/location.svg?react';
-import DropDownImg from '../../assets/icons/common/dropdown.svg';
+import DropDownImg from '../../assets/icons/common/arrow/downArrow.svg';
 import useDevice from "../../hooks/useDevice.js";
 import useDropdown from "../../hooks/useDropDown.js";
